backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic

sqlalchemy
uliweb-alembic
python-dotenv